refactor: finalize backend and clean up employee service logic
- Refactored employee service methods for clarity and consistency
- Finalized implementation of CRUD operations
- Added null checks and proper error handling using ResourceNotFoundException
- Improved update logic to safely map updated fields
- Ensured controller endpoints are clean and RESTful
- Backend now complete and ready for integration with frontend

+++ src/main/java/com/production_grade/employee_management/service/EmployeeService.java
@@ -1,14 +1,14 @@
 package com.production_grade.employee_management.service;
 
 import com.production_grade.employee_management.entity.Employees;
+import com.production_grade.employee_management.exception.ResourceNotFoundException;
 import com.production_grade.employee_management.repository.EmployeeRepository;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
-
+//thyis class implements the EmployeeServiceInterface with crud methods
 @Service
-public class EmployeeService {
+public class EmployeeService implements EmployeeServiceI {
     private final EmployeeRepository employeeRepository;
 
     public EmployeeService(EmployeeRepository employeeRepository) {
@@ -17,21 +17,45 @@ public class EmployeeService {
 
 //    CRUD operations code below
 
-// create student
+// create employee
     public Employees createEmployee(Employees employee) {
         return employeeRepository.save(employee);
     }
 
-//    read all students
+//    read all employee
     public List<Employees> getAllEmployees() {
         return employeeRepository.findAll();
+    }
+
+//    read employee by id
+    public Employees getEmployeeById(Long id) {
+        return employeeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Employee with id " + id + " not found"));
     }
 
 //    update employee
 
 //    delete employee
-    public String deleteEmployee(Long id) {
+    public void deleteEmployee(Long id) {
+        if (!employeeRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Employee with id " + id + " not found");
+        }
+
         employeeRepository.deleteById(id);
-        return "Employee with id " + id + " deleted";
+
+    }
+
+//    update
+    public Employees updateEmployee(Long id,Employees updatedEmployee) {
+        Employees existingEmployee = employeeRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Employee with id " + id + " not found"));
+
+//        existing employee
+        existingEmployee.setEmployeeName(updatedEmployee.getEmployeeName());
+        existingEmployee.setEmployeeLastname(updatedEmployee.getEmployeeLastname());
+        existingEmployee.setEmployeeEmail(updatedEmployee.getEmployeeEmail());
+
+        return employeeRepository.save(existingEmployee);
+
     }
 }
